feat[php]: (#72) confirm passphrases

+++ php/mint-config/1.php
@@ -16,6 +16,7 @@
                 <div data-tooltip="Minter Address: The L2 wallet address or ENS or Account ID of the minter"><input required type="text" class="form wide" id="minter" name="minter" placeholder="Minter Address [Wallet Address, ENS, or Account ID]" /></div>
                 <div data-tooltip="Minter Private Key: The Loopring private key of the minter [DO NOT SHARE THIS INFO WITH ANYONE]"><input required type="password" class="form wide" id="private_key" name="private_key" placeholder="Minter Private Key" /></div>
                 <div data-tooltip="Config Passphrase: A passphrase to encrypt your private key with [DO NOT SHARE THIS INFO WITH ANYONE]"><input required type="password" class="form wide" id="secret" name="secret" placeholder="Config Passphrase" /></div>
+                <div data-tooltip="Confirm Passphrase: Re-type the passphrase you previously entered [DO NOT SHARE THIS INFO WITH ANYONE]"><input required type="password" class="form wide" id="pass_confirm" name="pass_confirm" placeholder="Confirm Passphrase" onpaste="return false;" /></div>
                 <div data-tooltip="Default Royalty Percentage: Percentage of the price of a sale that will go to the Royalty Address (LooPyGen generated collections override this percentage)">
                     <input required type="number" class="form wide" id="royalty_percentage" min="0" max="10" name="royalty_percentage" placeholder="Default Royalty Percentage: 0-10" />
                 </div>
@@ -41,22 +42,27 @@
         $minter = $_POST['minter'];
         $private_key = $_POST['private_key'];
         $secret = base64_encode($_POST['secret']);
+        $pass_confirm = base64_encode($_POST['pass_confirm']);
         $royalty_percentage = $_POST['royalty_percentage'];
         $nft_type = $_POST['nft_type'];
         $fee_token = $_POST['fee_token'];
 
-        $config_data = array("minter"=>$minter,
-                             "private_key"=>$private_key,
-                             "royalty_percentage"=>(int)$royalty_percentage,
-                             "nft_type"=>(int)$nft_type,
-                             "fee_token"=>(int)$fee_token);
+        if ($secret !== $pass_confirm) {
+            $code = "100";
+        } else {
+            $config_data = array("minter"=>$minter,
+                                 "private_key"=>$private_key,
+                                 "royalty_percentage"=>(int)$royalty_percentage,
+                                 "nft_type"=>(int)$nft_type,
+                                 "fee_token"=>(int)$fee_token);
 
-        $config_json = json_encode($config_data, JSON_UNESCAPED_SLASHES|JSON_PRETTY_PRINT);
-        file_put_contents($mint_config, $config_json);
+            $config_json = json_encode($config_data, JSON_UNESCAPED_SLASHES|JSON_PRETTY_PRINT);
+            file_put_contents($mint_config, $config_json);
 
-        # encrypt the config file
-        $command = "encrypt --mint --secret ${secret}";
-        exec($command, $output, $code);
+            # encrypt the config file
+            $command = "encrypt --mint --secret ${secret}";
+            exec($command, $output, $code);
+        }
 
         Redirect("/mint-config/finish?result=${code}", false);
     }
